Move admin tests to admin.py
